# optimize replication (eliminate ArrayEntryBuffer)