@akiradeveloper 2020-11-06
```
test bench_apply_1             ... bench: 101,149,196 ns/iter (+/- 652,848)
test bench_apply_16            ... bench:   2,678,458 ns/iter (+/- 462,677)
test bench_apply_4             ... bench:   1,476,530 ns/iter (+/- 371,923)
test bench_apply_64            ... bench:  10,406,150 ns/iter (+/- 3,504,094)
test bench_commit_1            ... bench: 101,181,125 ns/iter (+/- 663,898)
test bench_commit_16           ... bench:   2,749,949 ns/iter (+/- 545,734)
test bench_commit_4            ... bench:   1,396,401 ns/iter (+/- 518,522)
test bench_commit_64           ... bench:  10,541,175 ns/iter (+/- 3,864,175)
test bench_query_1             ... bench:     543,109 ns/iter (+/- 350,937)
test bench_query_16            ... bench:     567,319 ns/iter (+/- 266,295)
test bench_query_4             ... bench:     527,484 ns/iter (+/- 224,848)
test bench_query_64            ... bench:     620,291 ns/iter (+/- 431,924)
test test_commit_huge_16_mem   ... bench:   5,424,252 ns/iter (+/- 2,026,191)
test test_commit_huge_16_rocks ... bench:   7,151,830 ns/iter (+/- 3,726,307)
test test_commit_huge_1_mem    ... bench: 101,225,251 ns/iter (+/- 590,917)
test test_commit_huge_1_rocks  ... bench: 101,557,093 ns/iter (+/- 693,465)
test test_commit_huge_4_mem    ... bench:   2,362,714 ns/iter (+/- 680,366)
test test_commit_huge_4_rocks  ... bench:   3,390,545 ns/iter (+/- 1,655,397)
test test_commit_huge_64_mem   ... bench:  20,304,060 ns/iter (+/- 6,745,655)
test test_commit_huge_64_rocks ... bench:  27,272,589 ns/iter (+/- 19,271,983)
```